move test-only functions to test file, fix lint
