+++ app/gui2/tsconfig.app.vitest.json
@@ -2,9 +2,7 @@
   "extends": "./tsconfig.app.json",
   "exclude": [],
   "compilerOptions": {
-    "composite": true,
     "lib": [],
-    "outDir": "../../node_modules/.cache/tsc",
     "types": ["node", "jsdom", "vitest/importMeta"]
   }
 }
