+++ app/gui2/ydoc-server/__tests__/edits.test.ts
@@ -1,7 +1,56 @@
 import { fc, test } from '@fast-check/vitest'
+import { Position, TextEdit } from 'shared/languageServerTypes'
+import { inspect } from 'util'
 import { describe, expect } from 'vitest'
 import * as Y from 'yjs'
-import { applyDiffAsTextEdits, applyTextEdits, convertDeltaToTextEdits } from '../edits'
+import { applyDiffAsTextEdits, convertDeltaToTextEdits } from '../edits'
+
+// ======================
+// === Test utilities ===
+// ======================
+
+/** Apply text edits intended for language server to a given starting text. Used for verification
+ * during testing if generated edits were correct. This is intentionally a very simple, not
+ * performant implementation.
+ */
+export function applyTextEdits(content: string, edits: TextEdit[]): string {
+  return edits.reduce((c, edit) => {
+    try {
+      const startOffset = lsPositionToTextOffset(c, edit.range.start)
+      const endOffset = lsPositionToTextOffset(c, edit.range.end)
+      return c.slice(0, startOffset) + edit.text + c.slice(endOffset)
+    } catch (e) {
+      throw new Error(
+        `Failed to apply edit ${inspect(edit)} to content:\n${inspect(c)}\n${String(e)}`,
+      )
+    }
+  }, content)
+}
+
+function lsPositionToTextOffset(content: string, pos: Position): number {
+  const lineData = getNthLineLengthAndOffset(content, pos.line)
+  if (pos.character > lineData.length)
+    throw new Error(
+      `Character ${pos.character} is out of bounds for line ${pos.line}. ` +
+        `Line length ${lineData.length}.`,
+    )
+  return lineData.offset + pos.character
+}
+
+function getNthLineLengthAndOffset(content: string, line: number) {
+  const nthLineRegex = new RegExp(`((?:.*\\n){${line}})(.*)(?:\n|$)`)
+  const match = nthLineRegex.exec(content)
+  if (!match) throw new Error(`Line ${line} not found in content:\n${content}`)
+  return {
+    offset: match[1].length,
+    length: match[2].length,
+  }
+}
+
+// ==================
+// === Test suite ===
+// ==================
+
 describe('applyDiffAsTextEdits', () => {
   test('no change', () => {
     const edits = applyDiffAsTextEdits(0, 'abcd', 'abcd')
@@ -18,6 +67,7 @@ describe('applyDiffAsTextEdits', () => {
       },
     ])
   })
+
   test('two adds', () => {
     const before = 'abcd'
     const after = 'abefcdxy'
