+++ app/ide-desktop/lib/dashboard/package.json
@@ -5,7 +5,6 @@
   "private": true,
   "scripts": {
     "typecheck": "tsc",
-    "lint": "eslint src",
     "build": "tsx bundle.ts",
     "watch": "tsx watch.ts",
     "start": "tsx start.ts",
