+++ app/ide-desktop/lib/types/tsconfig.json
@@ -1,8 +1,10 @@
 {
+  "include": ["./*"],
   "compilerOptions": {
     "types": ["node"],
     "lib": ["ES2015", "DOM"],
-    "skipLibCheck": false
+    "skipLibCheck": false,
+    "outDir": "../../../../node_modules/.cache/tsc"
   },
   "extends": "../../tsconfig.json"
 }
